Invert dependency relationship of coyote-error, coyote-proto
