+++ crates/coyote-proto/src/lib.rs
@@ -1,9 +1,13 @@
+mod error;
 mod msgpack;
 mod msgpack_client;
 mod msgpack_or_json;
 pub mod prelude;
+mod validation;
 
 pub use self::{
+    error::StandardErrorBody,
     msgpack::MsgPack,
     msgpack_or_json::{MsgPackOrJson, capture_accept_hdr},
+    validation::{ValidationErrorBody, ValidationErrorItem, validation_error, validation_errors},
 };
