+++ crates/coyote-proto/src/validation.rs
